파일읽기 -> ArrayList로 변환하기 진행중(V1)

+++ src/main/java/com/ll/exam/Post.java
@@ -1,11 +1,11 @@
 package com.ll.exam;
 
 public class Post {
-    int num;
+    Long num;
     String author;
     String talk;
 
-    public Post(int num, String author, String talk) {
+    public Post(Long num, String author, String talk) {
         this.num = num;
         this.author = author;
         this.talk = talk;
